Improves UI and adds map track display
Refactors the UI for better usability by increasing the size of several panels and controls.

Adds functionality to display tracks on the map from selected log files.
Error handling and messages are included to notify when no map-related data is available.
A chart is added to the right panel, although it is not yet populated with data.

Also includes a minor fix to .gitignore to ignore the dist/ folder.